use pip and install_requires
Since rasterio 1.0 is released, we should be able to simply use pip for testing.

Add extras_require which allows us to customise the install to include dev or optional dependencies. This replaces the requirements.txt files.

Also switches the GitLab CI image we use to the more commonly used debian based python image.