deprecate idf.memmap and idf.dataarray
Fixes #35.
Did not yet remove the `memmap` keyword argument, but we just
give a deprecation warning and ignore `memmap`.
This can only possibly give issues if a single IDF is larger than RAM,
which is unlikely.
If the need for supporting larger-than-RAM IDFs exists, make an issue.
We can probably support it with chunks. Currently a single IDF is a
single chunk, but this can probably also be cut up.