Get slots working on python < 3.7
I'm still on Python 3.6.7, and getting the error message:
"in __slots__ conflicts with class variable"

This is apparently a difference between this python version and later
ones. Due to different Python versions (maybe xarray version also plays
a role?), CI does not pick this up.

At any rate, for the xarray subclasses, the template is no longer
initialized within the __init__, but before it. This seems fine, the
templates are simply loaded when the classes are defined, rather than at
class instantiation; is marginally more efficient as well (though not at
startup...).