Fix recursion error in array_io.writing
The issue was that a.isel({dim: slice(start, end)}) does nothing in reducing the dimensions with chunks. It will succesfully select a single chunk in this dimension, but if there is a second dimension with chunks, these will obviously still exist.

Concretely, say there's dims time, layer, y, x, with chunks:
((1, 1), (1, 1, 1), (1300,) (1200,))
The first isel will select have this as a result:
((1,), (1, 1, 1), (1300,) (1200,))
The first chunk, time, still exists, and max(map(len, a.chunks)) == 1 will evaluate False (since the biggest chunks has length 3)).

Closes #105.