Compute mxss in the same way as imod-wq; Write only necessary scalars.
iMOD-wq has some annoying inconsistenties when it comes to setting concentrations. It accurately checks the number of active elements when concentration is given by IDF, but not when given by scalar.

In case of a scalar, it allocates a FULL array, the entire domain. This means that crch_t1_p?_l? allocates nlayer * nrow * ncolumn elements. To avoid this, the methods figure out which layers require concentrations, and only writes those to the runfile.

MXSS is also computed accurately for scalar concentrations, by multiplying the number of layers by the full number of row and columns.

Fixes #119.