Add ClippedBoundaryConditionCreator class
This commit adds a ClippedBoundaryConditionCreator class. This class is used to create a ConstantHeadBoundary condition based on already existing ConstantHeadBoundary conditions. It determines which boundaries don't have a BC yet and assigns a ConstantHeadBoundary condition based on a `state_for_boundary` input

In #420 the ClippedBoundaryConditionCreator is hooked up to the main code