Preserve any unassigned boundary conditions of the original domain on the clipped domain
This commit handles the following edge case: If the original model has a boundary which doesn't have any boundary condition assigned to it then this state should also be preserverd on a boundary in the clipped model that coincides with that boundary.

To handle this situation we reuse the `__create_boundary_condition_for_unassigned_boundary`. When applied to the original model it will produce a constant head packages for boundariwa without a boundary conditions. When we call the `__create_boundary_condition_for_unassigned_boundary` for the clipped model we supply that created boundary condition. The `__create_boundary_condition_for_unassigned_boundary` will add it to a list of existing boundary conditions and therefor no boundary conditions will be created for it