Add support for unstructured grids to the well package
This commit adds support for unstructured grids to the well package. A ones_like method has been created that depending on the input grid calls forwards the call to xarray or xugrid
Further a GridDataArray typing has created in the grid file that can be referenced from other files.
Finally the test_write_well has been parameterized to test the wells with both structured as unstructured grids