Refactor MF6Well to adhere to the write_datafile method
In this commit the Mf6Wel has been refactored to adhere to the write_datafile pattern. The write_datafile method in the base class (BoundaryCondition) calls the \_ds_to_arrdict and \_to_sparse methods. The derived classes should override those methods if they want to implement custom behaviour.

The reason to make this change is two fold:

1. The \_to_sparse method defined in the Mf6Wel package doesn't implement the same method definition as is defined in the BoundarCondition class
2. The makes possible to make all high level packages consistent with eachother. For instance the HorizontalFlowBarrier only has to implement the \_to_sparse method. There is no need to override the write_datafile method