Add high level HorizontalFlowBarrier package
In this commit the grid agnostic HorizontalFlowBarrier package is added. The logic is defined in the abstract `HorizontalFlowBarrier` class. 3 specialization exist to help the user create the barrier. Foreach of those specialiatzation the user has to provide a GeoDataframe wich consists of:

* a geometry of linestrings that define the barrier
* the ztop to which the barrier reaches
* the zbottom to which the barrier reaches
* the values applied to the barrier

When a ztop or zbottom doesn't fully covers a layer an effective values is computed

The unit tests make use of patching framework to decouple the code. To make this work the __new__ method had to be added to he PackageBase class. This solution has been provided on stackoverflow: https://stackoverflow.com/questions/64737213/how-to-patch-the-new-method-of-a-class