Fix/Update the package_instance_creation and package_sanity tests. Cleanup hfb constructor
In this commit the `package_instance_creation` test and the `package_sanity` test are updated.

For the package_sanity test to succeed with the hfb package a separate `to_netcdf` method has been introduced in the pkgbase class. This method calls the protected method `_netcdf_encoding` which by default returns an empty dictionary. This method can be overridden in any of the derived packages. This is necessary for the hfb package because it has a LineString object in its dataset which cant be handled by xarray's `to_netcdf` method. By providing an encoding this object is new encoded as a string. Further more, to also be able to read the netcdf file, the `from_file` has been overridden in the hfb class. When processing the linestring it is converted back from a str to a shaply object

Other changes in this commit are:

* Adding the high level packages to the `package_instance_creation` tests
* Cleaning up the constructor of the hfb package, removing the unused top and bottom parameters
* Skipping xugrid version 0.6.3 because it contains a bug in the `snap_to_grid` method