Refactor hfb and the hfb_adapter class. Move all the grid specific knowledge from the adapter to the hfb class
In this commit the all domain specific knowledge has been removed from the `hfb_adapter` and moved to the the `hfb` package. A `to_connected_cells_dataset` method has been added that does the conversion from edge based values to connected cell values. The code works for both structured and unstructured grids.

The `hfb` creation has temporarily been disabled in the `disv converter`. It will be reenabled after #513

Further more the computation of the barrier multiplier has been updated. There was a bug in the computation where it didn't honored the provided `ztop` and `zbottom` values

Lastly validation has been added to identify valid edges. An edge for a barrier is considered valid when it is not on the domain boundary or is not connected to an inactive cell