Use new HorizontalFlowBarrier package in the disv_converter
In this commit the disv_converter is updated to use the new HorizontalFlowBarrier package The disv_converter assumes that the barriers defined in the prj file are resistance flow barriers

The imod_conversion.py has been updated as well. It now no longer ignores the barrier in the project files

Finally to a inactive cells mask has been added and applies in the hfb package. This is needed because the barrier imported from the prj file used in the imod_conversion.py intersect with inactive cells. The parts of the barrier that are adjacent to an inactive cell are ignored to be compatible with MODFLOW