Create grid agnostic HorizontalFlowBarrier package
This commit adds the new HFB package. A flow barrier can be created by creating a HorizontalFlowBarrierHydraulicCharacteristic, HorizontalFlowBarrierMultiplier or HorizontalFlowBarrierResistance. A geopandas Dataframe needs to be provided which contains the geometry of the barrier, the value of the barrier and the z-top and z-bottom values of the barrier. The old HFB class has been moved to the mf6 mf6_gfb_adapter file. The code has been refactored in such a way that the adapter has no knwodledge about the domain.

Futher:

* has the `_to_sparse` method been renamed to `_to_struct_array`.
* has the disv converter been updated to use the new HFB package
* has the to_netcdf been added to the pkgbase to allow more control over the serialization process