Add MyPy reporting to the CI pipeline
At the moment the MyPy reporting is limited to the imod/mf6 folder. The pipeline stage will not produce an error only a warning when it fails. When all the MyPy errors have been fixed this can be made more strict, failing whenever MyPy throws an error

In the future more folders can be added to the MyPy analyzing stage