Gitlab #510 Add split method to the simulation class. Splits all the models in a simulation using the provided subdomain_labels.
In this commit the split method is added to the simulation class. The method accepts an submodel_labels array which is used to split the domain.
The method returns a new simulation with the postfix partioned added to it. The models in the new simulation have the original model name with a partition index added to them.
The modelsplitter file has been refactored a bit to be usable for the split method. The corresponding test file was named incorrectly and has been renamed.
In the oc class an extra check has been added for the save_concentration variable. It could happen that save_concentration is of type DataArray and don't contain any items. In that case save_concentration is set to None