Gitlab #529 Deactive cells that don't belong in a sliced submodel
When a structured grid is sliced, it may still contain cells that don't belong to the active domain.
In this commit those cells are deactivated.
The create_domain_slices method has been renamed to create_partition_info and now provides a list of SubmodelPartitionInfo. This datastruct consists next to the domain slices also the label_id and the active idomain. The latter is used in the _slice_package method to deactivate cells that don't belong in the sliced domain.
Further more the unit tests have been cleaned and the unit test for the structured grid as been updated