Gitlab #521 Remove packages that become invalid after splitting.
This commit checks if a package is valid after splitting it. If it is invalid it isn't added to the split model.

The split_package method has been replaced by a more general clip_by_grid method which works for structured/unstructured and for regular and grid agnostic packages. Interfaces have been introduced to make this work.

Because of the latter change the modelsplitter file has been reduced significantly