Gitlab #547 Add temporary partitioned TWRI model for testing purposes
This commit adds a copy of the `ex01_twri` with the addtition of a call to the split method. The user can defined the number of partiones by changing the `number_partitions` variable. np.geomspace is used to create the partitions. This has been used in favor of np.linspace because the resulting partitions will have different sizes and therefor different exchange files. This contributes to a more thorough testing case.

This file has been added tot aid in testing purposes and should not be merged to the master branch. Issue #548 has been created to remove this file and in issue #324 a proper example for the user will be created.

Further more this commit fixes a bug in splitting the well package. The splitted well package can exists of zero wells. THose kind of packages should be invalid and therefor be removed. The `_write_schemata` of the well package has been updated to reflect that condition