Gitlab #539 Create GWF-GWF exchanges for structured grids
In this commit the GWF-GWF exchanges are created for a simulation with a structured grid that is partitioned using a domain_label array. A separate `ExchangeCreator` class has been created to do so. In this class xarray's diff method is used to find the boundaries of the subdomain. With this information a connected cells dataframe can be created. The dataframe contains information about the connected cells in terms of the global grid. Therefor a mapping between the global grid cells and local grid cells also are being determined. With those 2 inplace the derivation of the exchanges can be created.

This commit does not write the exchange files. This will be done in #540 This commit does not compute the geometric coefficients. That will be done in #325 This commit only works for structured grids. Unstructured support will be added in #541