Gitlab #325 ﻿Compute geometric information of connected cells on a structured grids
This commit adds the geometric information (`cl1`, `cl2`, `hwva`) to the exchange files of simulations with a structured grid. The computation is done in the `_compute_geometric_information` method in the `exchange_creator` class. It relies on the `geometric_grid_info` which contains the `x`- and `y`-coordinates and the `dx` and `dy` sizes of a grid

The `to_cell_idx` has been moved to the `grid_utils` file

Furthermore the `parameterizable_basic_dis` has been extended to add more flexibility to the generated domain. We are now able to generate a non-equidistant grid that we used to test the computed geometric information