Gitlab #550 Refactor data format gwfgwf
In this commit the GWFGWF exchange class has been cleaned up. This was able by adjusting the jijna template of the exchange file a little bit. By doing so the render method of Base class could be used instead of implementing our own. The cell index variable had to slighlty change as well. It now consists out of an array of array instead of an array of tuples. This also allows it to work with the netcdf dump

Further more this commit consists out of a fix of the gwfgwf tests. They weren't discovered (and therefor weren't run) because of an incorrect naming of the test class. Capitalizing the test class name resolved this