Gitlab #567 Fix failing plot_map.py example
Both locally and on the Gitlab servers the plot_map.py test is failing. It became clear on the CI first because sphinx wasn't able to generate images.

For the latest release of contextily (i.e. 1.4.0) it is mentioned that the default tile provider Stamen is no longer free to use and therefor they are switching their default provider to OpenStreetMap. Doing the same in our code resolves the issue.