Gitlab #568 Only build the documentation when merging to master
This commit restricts the building of the documentation to MR's to the master branch. This will avoid that the documentation is build on story branches which costs precious CI minutes. The documentation is only published on the master branch.

The reason we still want to build the documentation before merging to master is that the examples are being run as part of the build stage. The examples are basically integration tests so we want to them to succeed before merging to master

Further the run of the unit tests has been optimized. The units tests were run twice (and some tests even thrice). This commit fixces that and the unit tests are only run once now.

A last optimization is that the imod-environment file has been split in a default and dev file, and that the modflow 6 packet is downloaded using conda

The total time is the same but on commit that don't merge to the master branch the documentation stage isn't called which should save us 10 minutes CI time

As an extra the pytest coverage results are now also reported on gitlab