Gitlab #538 Create and write the MultiModel exchange files
MultiModel support is added. The user can create such a model by calling the split method on a simulation object. A submodel labels array needs to be provided which indicates to which partition a cell belongs. The split method creates a new simulation in which the original models are split. The new simulation also contains gwf-gwf exhcanges packages which defines the connection between the partitioned submodels. This works for both structured and unstructured grids