Gitlab #584 Write auxiliary variables to the exchange files
In this commit the auxiliary variables are written to the exchange files. This has been accomplished by updating the corresponding jinja file and by adding the auxiliary variables as a dict to the input dict of the render method

Further the alignment of the exchangedata has been corrected. The are now indented by 2 spaces