Gitlab #579 Add auxiliary variables to the exchanges files
This commit contains a refactor of the existing auxiliary variables logic. It has been simplified an extracted to a separate file so it can be reused by other packages then only the stress packages. This new extracted logix is being applied to the gwfgwf exchange class, enabling the angldegx and cdist variables to be written to the exchange files