Issue #677 Transition CI build to use Pixi (#716)
This commit adds the Pixi build system to the project. It enables us to
simplify the CI pipeline, speed up the pipeline, debug the pipeline
steps locally and install a developer environment with more ease.


The imod-environment.yml and imod-environment-dev.yml files are no
longer required and have been removed