Issue #709 fix hfb structured grids (#710)
Closes #709 and #659

This PR implements a fix for:

* HorizontalFlowBarrier lines not being properly clipped in
clip_by_grid, due to bounding_polygon also returning a polygon for
active values.
* When the barrier is placed at the edge of a model partition it is
removed. This can create empty HFB packages which should be removed,
this now raises an error.
* The clip_by_grid can introduce MultiPoints and MultiLineStrings for
ILineDataPackage. The MultiPoints are thrown out, the MultiLineStrings
are separated into separate LineStrings.

UPDATE 03/01/2024:
after comment
https://github.com/Deltares/imod-python/pull/710/files/4df2e20c82b07c111890cd85d85696924da3296b#r1439526060
some code was refactored into a separate property ``line_data``. This
has the advantage that it is now more accessible to access line data as,
and set it with, GeoDataFrames.