Issue #672 remove coupling logic (#727)
Closes #672 

Removes coupling logic. This logic has not been used a lot that I know
off, but it now throws an error if somebody tries to initialize the
MetaMod class. The example for the docs has also been updated to point
to the new location of the example [(in
iMOD-Documentation)](https://github.com/Deltares/imod_coupler/tree/main/tests/test_primod).

I think this is sufficient (instead of a grace period where
``DeprecationWarning``s are thrown), as it hasn't been applied too much
yet: I know of somebody experimenting with it and mostly the iMOD
Coupler test bench, which already moved.

``primod`` is going to be used this year, so good to have to point
people in the right direction.