Issue #622 properly compute aux vars cdist and anglex for unstructured grids (#723)
This PR introduces tests for the geometric coefficient angledegx which
is computed at partition boundaries. Angledegx is used for specific
discharge calculations, hence the testing for specific discharge results
before and after splitting a model. An example was also added showing
how both flopy and imod-python can be used to do the model splitting.

---------

Co-authored-by: Sunny Titus <sunny.titus@deltares.nl>