Issue #535 open hds convert time (#749)
This PR adds the possibility to convert the time coordinates in datasets
obtained from MF6 output ( heads, in this PR) into calendar time. To do
so, the user must add arguments to the open_hds function specifying the
calendar time and date of the start of the simulation, and the internal
time unit used in the model (days? seconds? etc)

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>