Issue #737 Add PR template (#758)
Fixes #737 

This PR adds a template for PRs. [Unfortunately the template only works
if it is in the default
branch](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/about-issue-and-pull-request-templates#pull-request-templates),
(``master``) in our case, so I could not test it. I presume it works
though, [as it works for
Ribasim](https://github.com/Deltares/Ribasim/blob/main/.github/pull_request_template.md).

[Relevant
docs](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/creating-a-pull-request-template-for-your-repository)