#751 Create imod logger (#754)
This commit adds the imod.logging.logger to the imod-python project.
It abstract away the use of external logger dependencies, making sure we
are not strongly coupled to external logger frameworks, making it easy
to switch from logger in the future.
Also because it is based on an interface it is easy mockable in unit
tests( as an alternative to mocks you can also use the provided
NullLogger).

The logger is disabled by default. To activate it you call the
imod.logging.configure() method

3 loggers are available:
- NullLogger 
- PythonLogger
- LoguruLogger

Users can choose which one they want to use. If a user is using logging
themselfs, this makes it easy to integrate imod logging into that.

For the user that don't want to bother setting up logging them self
there is already a default stream handler initialized when the configure
method is called. A flag can be used to also enable outputting to a file

Examples on how to use the code are provided in the docstring in the
__init__ file