Issue #740 Fix incorrect arguments provided to the clip_box method of the well package (#741)
In this commit the wrong arguments provided to the well package is
addressed.

The Well package is a grid agnostic package. That means that the
layer_min and layer_max arguments from the clip_box method has no
meaning to it.

To still conform to this method signature the top and bottom of a grid
is provided. Those are then used to retrieve a z_min and z_max