Issue #748 Fix failing transport model examples i.c.w. the MODFLOW6 nightly build (#771)
Fixes #748

# Description
The unit tests and examples were failing when run against the latest
MODFLOW6 nightly. This is caused by a change on the MODFLOW6 side were
any gwf-gwt exchanges has to have an actual existing file (even though
it is emty).
The precious behaviour was that just mentioning the gwf-gwt exchanges in
the nam file was enough.

This commit adds the gwfgwt package. It uses the same logic as the
gwfgwf exchanges in terms of how it is written to disk.
Any common methods between the different type of exchanges has been
moved to the newly created ExchangeBase class

**Note**
I didn't add any unit tests for the simulation class. The reason for
this is twofold:
- I moved code around, no major functional changes
- We have to refactor they way the gwf-gwt exchanges are being created
the moment we start working on the multimodel support for gwt, making
any unit tests I add obsolete

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example