Use hatchling instead of setuptools, add pixi task: pypi-publish (#760)
This setups the pyproject to use hatching instead of setuptools.

It also adds a pypi-publish task to pixi to make the release process a
little smoother.

Finally, the pyproject version is derived from the imod.__init__.
Ideally, there's a single source of truth, but pixi still contains its
own version number (note: that one is not used anywhere, you could set
it to a dummy value as well).

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>