Issue #757 Create nightly pipeline (#777)
Fixes #757

# Description
This commit adds a nightly pipeline.
The goal of the nightly pipeline is to build or repository against the
latest development branch of MODFLOW.
This will allow us to catch breaking changes early on.

The pipeline file has been refactored to prevent code duplication. This
has been accomplished by using templates.
The new nightly pipeline is added as a new subproject of the main
project

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example