Issue #708 Refactor dataset initialization (#722)
Fix https://github.com/Deltares/imod-python/issues/708

The changeset is incomplete, as it is not rolled out for all packages
yet. Therefore tests will fail, hence the draft status. This to make the
review process more focused, reviewers are asked to provide feedback on
the approach.

Changes:
- Variable names are inferred based on what is provided to locals().
This makes it possible to avoid having to manually assign variables in
the __init__ of each specific package.
- I had to add one function to work around this issue:
https://github.com/Deltares/xugrid/issues/179
- Grids are merged with an exact join, this to avoid that xarray joins
variable coordinates in an unwanted way when dataarrays are
inconsistent, which causes issues like
https://github.com/Deltares/imod-python/issues/674
- Refactored the riv unit tests to use pytest cases, so that tests
(without the xr.merge) are run for both unstructured and structured
data.

Update:
- Variable names are not inferred with locals() anymore, but instead
with a ``pkg_init`` decorator.