Issue #783 Fix all remaining MyPy errors (#784)
Fixes #783

This commit fixes all remaining MyPy errors in the mf6 module

The PR is split in several commits, each commit contains a certain file
in which the mypy errors are fixes or exist of a type of error that
crosses multiple files. I suggest reviewing this PR commit by commit.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example