Issue #809 Update project dependencies (#810)
Fixes #809

This commit updates the dependencies of the project.
There are some issues with the latest dask packag,s which isn;t playing
nice with raterio, and with the pytest package, which has issues with
pytest-cases. I've pinned both packages to a version that doesn't cause
any issue.
Further there has been an update of black which results in a new
formatting rule being applied.

Hopefully this is the last time we have to update manually as the
pipeline script to do this automatically is almost done

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example