Issue #795 Add pipeline which periodically updates the projects dependencies to the latest version (#804)
Fixes #795

This commit adds a pipeline that periodically updates the project
dependencies.
It runs every sunday at 15:00. When changes are found a new update
branch is pushed and a PR is created.

The created PR in turn will trigger the normal build pipeline so all the
regular checks are run against this branch.
If any of the tests fail they should be fixes as part of the created PR

Reviewers are automatically assigned in this case that are @luitjansl
and @JoerivanEngelen

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example