Issue #765 gwt gwt exchanges (#800)
Fixes #765 

# Description
creates gwtgwt exchange packages when splitting a simulation that
contains 1 or more transport models. A gwtgwt package is created per
partition boundary and per transport model. (while the geometry
constants are the same for all transport models, the options in the
gwtgwt file are specific to single transport models. However, for now,
these settings are not set; this will be the topic of #747
Also adds tests.

 Checklist

- [X] Links to correct issue
- [ ] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>