Issue #816 Add MyPy static code analyze step to the pipeline (#817)
Fixes #816 

This commit adds a MyPy build step to the build pipeline.
It will run as part of the regular build as well for the nightly build
Both the mypy_report as the regular mypy commands are called. The first
command produces a report that can be displayed in teamcity the same way
unit tests are. The second command will also give an output to the build
log

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example