Issue #826 Remove black in favor of ruff (#827)
Fixes #826

This commit replaces the black linter/formatter in favor of the ruff
linter/formatter.
The ruff and black formatter are 99% compatible but they vary just
slightly.
This causes issues when calling the `pixi run format` task because they
counteract each other.

A full list of differences between black and ruff can be found at:
https://docs.astral.sh/ruff/formatter/black/

The changes to review are in the pixi.toml file. All other changes are
made by the ruff formatter.

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example