Issue #836 Make MyPy stricter by enabling follow_imports (#837)
Fixes #836

In this commit the settings of MyPy are made stricter by enabling the
follow_imports flag. By enabling this flag, information about type hints
from modules being imported are also taken into account when performing
the static code analyzes

In the pipeline a threshold has been added to the MyPy step which
tolerates 2 errors. These 2 errors have to do with mypy errors with
regard to the regridder. Those errors are better handled in a separate
story. Once resolved the threshold can be set to 0

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example