Issue #851 Pixi debug env (#847)
Fixes #851

# Description
<!---
Thanks for opening a PR!

Please add your description here of changes made and how they are going
to
resolve the linked issue 
-->
Since the latest pixi release, support has been built in for multi
environments, [as described
here](https://pixi.sh/design_proposals/multi_environment_proposal/#motivating-example).

This changeset adds an environment for local interactive python. This
makes it a lot easier to make plots in VSCode, as you can produce them
in the ipykernel. Without this, plots need to be made in the debugger,
which is notoriously annoying to get to work with Matplotlib, saved to
disk.

To install:
``pixi run --environment interactive install ``
``pixi shell --environment interactive``

Now open VSCode, select the python.exe in .pixi/envs/interactive as
interpreter, create a script, press CTRL+A to select all and press
SHIFT+ENTER to run the code in the interactive kernel.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [ ] Links to correct issue
- [ ] Update changelog, if changes affect users
- [ ] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example