Issue #601 deepcopy (#866)
Fixes #601

# Description
Fixes deepcopying for model and simulation objects. 
The issue was that the jinja-template was an instance variable in these
classes, and when making a deepcopy python attempted to copy them. But
the jinja templates are not copyable, so this caused a crash. Now the
templates are class variables for model and simulation, so when making a
deepcopy, the same template object is used for the copy as for the
source.

# Checklist

- [X] Links to correct issue
- [ ] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example