Issue #807 split twice (#922)
Fixes #807 

# Description
The issue itself was a bugreport, it said that after splitting a model,
calling "split" a second time on the input model would result in a
crash. This could not be reproduced. Still, a test was added to make
sure we can split a model twice in the same way and that we get the same
result in both cases.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [X] Links to correct issue
- [ ] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example