Issue #934 Fix PiXi linting and formatting commands (#935)
Fixes #934 

# Description
The lint and formatting commands as defined in the pixi.toml file are
incomplete.
The ruff linter is used in the linting task while the ruff formater us
used in the formatting task
However the ruff linter checks for different issues then the ruff
formatter does.

To fix this the linting tasks has been extended to use `ruff format
--check` while the formatting tasks has been extended to use `ruff check
--fix`. Now the linting and formatting tasks should not have any
discrepancies anymore.

Further more the format commands has been applied. This resulted in 100
files being formatted. These changes can be found in the second commit.

# Checklist
- [x] Links to correct issue
- [ ] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [ ] Unit tests were added
- [ ] **If feature added**: Added/extended example