Issue #429 regridding cache (#933)
Fixes #429

# Description
Now caches weight sets instead of regridders when regridding models.
This saves re-computing weights, because the same weights are used for
different regridders (if the regridder type is the same and only the
regridding method is different).
For example if we have 2 overlap regridders, but one is of method "mean"
and the other one is of method "geometric_mean" they will use the same
weights.

# Checklist

- [X] Links to correct issue
- [X] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example

---------

Co-authored-by: Joeri van Engelen <joerivanengelen@hotmail.com>