Issue #536 misleading well message (#957)
Fixes #536 

# Description
This PR solves a bug. When a well was added outside of the domain, this
resulted in an error message stating that the
well was not added to the model because it violated constraints of
minimum thickness / minimum k. The real reason, that the well was
outside of the domain, was not mentioned.
Now a separate check has been added that checks if the wells are inside
the domain, and if not it give an error message
stating this, as well as the coordinates of the first well outside of
the domain.

#d Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [X] Links to correct issue
- [X] Update changelog, if changes affect users
- [X] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [X] Unit tests were added
- [ ] **If feature added**: Added/extended example