Issue #950 Distribute conductances (#973)
Fixes #950

# Description
Adds the following things:
* Adds utility functions to distribute conductances over depth, exposed
to public API
* Moves allocation tests to test_topsystem.py file, so cases can be
shared with the distribute conductances test.
* Moves cases to separate source file, do not use objects anymore to
group cases, but function name prefixes. Furthermore tags are used to
use tag which cases are used for RIV, DRN, GHB, and RCH tests. This
saves quite some duplicate code.
* Move code ``basic_dis`` fixture to function ``make_basic_dis``, which
is used to also make a 4 layered discretization fixture.
* Move duplicate code in allocation source file to
``_enforce_layered_top``

The only thing still missing is some extra options to distribute
conductances for the Drain package. In this PR, it is not possible to
use the crosscut thicknesses and correction factors yet. I'll make a
separate task for that to not make the PR too overwhelming.

# Checklist
<!---
Before requesting review, please go through this checklist:
-->

- [x] Links to correct issue
- [x] Update changelog, if changes affect users
- [x] PR title starts with ``Issue #nr``, e.g. ``Issue #737``
- [x] Unit tests were added
- [ ] **If feature added**: Added/extended example